Diag should log to console if it is requested or diag service is not available
