+++ index.coffee
@@ -1,7 +1,7 @@
 udp         = require 'dgram'
 async       = require 'async'
 Protocol    = require './protocol'
-log         = require './diag'
+log         = require './log'
 zmq         = require 'zmq'
 Constants   = require './constants'
 Convert     = require './convert'
