+++ lib/diag.js
@@ -1,28 +1,10 @@
-var Diag, Log, Protocol, Variable, cliOptions, os, util,
-  __slice = [].slice;
+var Diag, Protocol, cliOptions, os;
 
 cliOptions = require('nomnom').parse();
 
 Protocol = require('./protocol');
 
 os = require('os');
-
-util = require('util');
-
-Variable = (function() {
-  function Variable(content) {
-    this.content = content;
-  }
-
-  Variable.prototype.toString = function() {
-    return util.inspect(this.content, {
-      depth: null
-    });
-  };
-
-  return Variable;
-
-})();
 
 Date.prototype.yyyymmdd = function() {
   var dd, mm, yyyy;
@@ -38,10 +20,6 @@ Date.prototype.hhmmss = function() {
   mm = ('0' + this.getMinutes().toString()).slice(-2);
   ss = ('0' + this.getSeconds().toString()).slice(-2);
   return hh + mm + ss;
-};
-
-String.prototype.startsWith = function(other) {
-  return this.substring(0, other.length) === other;
 };
 
 Object.defineProperty(global, "__stack", {
@@ -99,6 +77,8 @@ Diag = (function() {
   Diag._headers = {};
 
   Diag._newHeaders = [];
+
+  Diag._isConsoleLogEnabled = false;
 
   Diag.startDate = function() {
     if (Diag._startDate == null) {
@@ -261,7 +241,16 @@ Diag = (function() {
     return (ellapsed[0] * 1e9 + ellapsed[1]) / 1000;
   };
 
-  Diag._log = function(level, args) {
+  Diag._createConsoleLogMessage = function(args) {
+    var argTexts, message;
+    message = "";
+    argTexts = args.map(function(val) {
+      return val.toString();
+    });
+    return argTexts.join(" ");
+  };
+
+  Diag._createDiagServiceMessage = function(level, args) {
     var argument, record, type, _i, _len;
     record = {};
     record.Process = Diag._getProcessInfo();
@@ -283,87 +272,21 @@ Diag = (function() {
     }
     record.Symbols = Diag._getNewSymbols();
     record.Headers = Diag._getNewHeaders();
-    return Protocol.sendDiag(record);
+    return record;
+  };
+
+  Diag._log = function(level, args) {
+    var record;
+    record = Diag._createDiagServiceMessage(level, args);
+    if (!Protocol.sendDiag(record || Diag._isConsoleLogEnabled)) {
+      return console.error(level + ": " + Diag._createConsoleLogMessage(args));
+    }
   };
 
   return Diag;
 
 })();
 
-Log = (function() {
-  function Log() {}
-
-  Log.levels = {
-    SILENT: 'silent',
-    TRACE: 'trace',
-    DEBUG: 'debug',
-    INFO: 'info',
-    WARN: 'warn',
-    ERROR: 'error'
-  };
-
-  Log.level = 'trace';
-
-  Log.trace = function() {
-    var args, _ref;
-    args = 1 <= arguments.length ? __slice.call(arguments, 0) : [];
-    if ((_ref = this.level) === 'trace') {
-      return Diag._log('VIRTDB_SIMPLE_TRACE', args);
-    }
-  };
-
-  Log.debug = function() {
-    var args, _ref;
-    args = 1 <= arguments.length ? __slice.call(arguments, 0) : [];
-    if ((_ref = this.level) === 'trace' || _ref === 'debug') {
-      return Diag._log('VIRTDB_SIMPLE_TRACE', args);
-    }
-  };
-
-  Log.info = function() {
-    var args, _ref;
-    args = 1 <= arguments.length ? __slice.call(arguments, 0) : [];
-    if ((_ref = this.level) === 'trace' || _ref === 'debug' || _ref === 'info') {
-      return Diag._log('VIRTDB_INFO', args);
-    }
-  };
-
-  Log.warn = function() {
-    var args, _ref;
-    args = 1 <= arguments.length ? __slice.call(arguments, 0) : [];
-    if ((_ref = this.level) === 'trace' || _ref === 'debug' || _ref === 'info' || _ref === 'warn') {
-      return Diag._log('VIRTDB_INFO', args);
-    }
-  };
-
-  Log.error = function() {
-    var args, _ref;
-    args = 1 <= arguments.length ? __slice.call(arguments, 0) : [];
-    if ((_ref = this.level) === 'trace' || _ref === 'debug' || _ref === 'info' || _ref === 'warn' || _ref === 'error') {
-      return Diag._log('VIRTDB_ERROR', args);
-    }
-  };
-
-  Log.setLevel = function(level) {
-    return Log.level = typeof level.toLowerCase === "function" ? level.toLowerCase() : void 0;
-  };
-
-  Log.enableAll = function() {
-    return Log.setLevel(Log.levels.TRACE);
-  };
-
-  Log.disableAll = function() {
-    return Log.setLevel(Log.levels.SILENT);
-  };
-
-  Log.Variable = function(param) {
-    return new Variable(param);
-  };
-
-  return Log;
-
-})();
-
-module.exports = Log;
+module.exports = Diag;
 
 //# sourceMappingURL=diag.js.map
