+++ lib/index.js
@@ -6,7 +6,7 @@ async = require('async');
 
 Protocol = require('./protocol');
 
-log = require('./diag');
+log = require('./log');
 
 zmq = require('zmq');
