+++ lib/protocol.js
@@ -73,31 +73,16 @@ Protocol = (function() {
     };
   };
 
-  Protocol.onDiagSocket = function(callback) {
-    if (Protocol.diag_socket != null) {
-      return callback();
-    } else {
-      return async.retry(5, function(retry_callback, results) {
-        return setTimeout(function() {
-          var err;
-          err = null;
-          if (Protocol.diag_socket == null) {
-            err = "diag_socket is not set yet";
-          }
-          return retry_callback(err, Protocol.diag_socket);
-        }, 50);
-      }, function() {
-        if (Protocol.diag_socket != null) {
-          return callback();
-        }
-      });
-    }
-  };
-
   Protocol.sendDiag = function(logRecord) {
-    return Protocol.onDiagSocket(function() {
-      return Protocol.diag_socket.send(proto_diag.serialize(logRecord, "virtdb.interface.pb.LogRecord"));
-    });
+    var ex;
+    try {
+      Protocol.diag_socket.send(proto_diag.serialize(logRecord, "virtdb.interface.pb.LogRecord"));
+    } catch (_error) {
+      ex = _error;
+      console.error("Couldn't send log message: ", ex);
+      return false;
+    }
+    return true;
   };
 
   Protocol.connectToDiag = function(addresses) {
